Create MOC_Code.m
Executive Summary 
Within this milestone, we are analyzing trash accumulation in the Atlantic Ocean by 
interpreting fluid motion in the ocean using different mechanics to calculate it. Studying these data 
is essential as studying waste disposal can get information about garbage and its longevity within 
ocean surface currents and potential for damage. They can study different products that do not 
break down after years in water and learn how ocean surface currents cause this build up of trash 
within the oceans affecting wildlife and ecosystems of the various species. The purpose of my 
study is to analyze ocean surface currents and calculate trash accumulation within the atlantic 
ocean based on these ocean surface currents data. 
Some of the mathematical techniques utilized within this study are partial derivatives 
(vector calculus operators), curl, divergence, scalar and vector fields. These were needed in order 
to understand and calculate fluid motion, flux, and the fields in which the ocean surface currents 
were moving according to the data we used to calculate trash accumulation within these ocean 
surface currents in the Atlantic Ocean. Divergence refers to the change in fluid motion based on a 
vector field, where if the divergence value is less than 0, it is a sink in the fluid. Negative 
divergence at a point means trash would accumulate inward because nothing would flow away 
from that point indicating that trash would build up there, hence the most probable location of trash
accumulation in the Atlantic Ocean are areas of negative curl, negative divergence and lower 
velocity. Just as we learned in physics about electric flux, the divergence of a flux in ocean surface 
currents results in the accumulation of trash, whereas if the divergence is negative, it is the loss of 
trash. Areas of lower velocity indicate that the currents are not strong enough to push trash away 
from each other causing more trash to accumulate. Within the ocean surface currents, only 
negative curl results in vortexes which is where the water becomes a sinkhole causing the trash to 
circulate in the water currents, whereas positive curl only circulates the trash. Hence, in order to 
determine areas of probable trash accumulation, we must seek areas that have negative curl, 
negative divergence, and lower velocity. 
In terms of the precision of my data and analysis, my calculations were accurate and of 
great quality as the calculations were clear and the trash accumulation for the area of interest had 
the greatest influence. It is good to refer to the characteristics of my data because it greatly 
compared to my intuition for the predicted location of the GAGP. The assumptions I made for the 
GAGP were considering areas of great trash build-up from currents of lower velocity and negative 
divergence and curl arguing the rotations of the water. My intuition can be reasonably compared to
the characteristics of my data proving the accuracy of my data source and analysis techniques as I 
used these physical properties to analyze and calculate these values about the currents. One 
improvement to improve the accuracy and meaningfulness of my study is to incorporate smaller 
values of contour to find more accurate current values than the values I calculated. 
Fig. 1 The figure shows the entire Atlantic Ocean with a shape outline that sketches the predicted 
location of the GAGP, i.e area of trash for accumulation in the ocean surface currents. View 
shows contour plot with axes label illustrating location of trash build-up and land boundaries. The 
plot has a color bar which indicates the surface current speed. 
Fig 2. Entire Atlantic Ocean view that shows ocean surface currents on the date of data given 
March 18, 2018. View shows a contour plot with land boundaries and the direction of the ocean 
surface currents. The plot has a color bar which indicates the surface current speed. 
Fig 3. The plot of the zoomed view of the area of interest’s current curl with the velocity vectors 
laid on top. The quiver arrows indicate that the area of interest has a negative curl. 
Fig 4. The plot of the zoomed view of the area of interest’s current’s divergence with the velocity 
vectors laid on top. The quiver arrows indicate that the area of interest has a negative divergence.
The plot has a color bar which indicates the surface current speed. 
Fig 5. The plot of the zoomed view of the area of interest’s current’s speed with the velocity 
vectors laid on top. The quiver arrows indicate that the area of interest has a negative divergence.
A zoomed out version was utilized in order to maximize the ocean surface currents speed into the 
area of interest and the area of interest’s velocity as the currents were barely visible. The plot has 
a color bar which indicates the surface current speed.